requirements.txt
numpy==1.19.2
PyYAML==5.3.1
